Multi Object detection sample crash when permissions are not granted
Summary:
MultiObject detection component "SentisObjectDetectedUiManager" is not checking if the WebCamTexture is ready during the update logic.

It causes a crahs when the permissions are not granted.

Under Unity 2022 the app is not crashing but you can see a bunch of errors.

Under Unity 6 the app finally crashes because of this error.

### Test apk:
https://drive.google.com/file/d/1px2onOabBXQkcvoTHkAjv7FFEHxblcZc/view?usp=drive_link

Reviewed By: kyrylokuzyk-m

Differential Revision: D71808535

fbshipit-source-id: 928d2716831c5c44da31ad9c1d6f6630b8ba8ede