Updated MultiObject detection bounding box placement
Summary:
Updated how we place the bounding boxes in front of the camera.

### 2D Canvas

Now, the 2D canvas that conatins the bounding boxes are placed when we draw the bounding boxes. It's position is the start inference position ( we capture the position before run the inference).

Instead of updating the canvas position every time, we update it when the inferences finishes, so the canvas reimains in the 3d space until the next inference.

This new behaviour removes the missplacing of the bounding boxes when the user was moving the head.

### 2D bounding boxes

Now the position and rotation of the bouding boxes are better calculated using Depth Api and PassthroughCameraUtils.GetCameraPoseInWorld().

### 3D Markers

Updated 3d marker placemet using camera instrinsics instead of  Unity world position transform.

Reviewed By: kyrylokuzyk-m

Differential Revision: D71739709

fbshipit-source-id: bb28ee628c3c0c3191f4e0510dbf014a906d60d6